Do not require tools to be installed in the same virtual environment
This is not required for `semgrep` and `mypy` -- the former cannot be
invoked with `python -m`. In that regard, the advice in the README was
misleading. There is no reason to require the tools to beinstalled in
the same virtual environment. The requirement makes this tool awkward to
use with `pipx` or `uvx`.

For all these reasons, this change removes that restriction for `fixit`,
`flake8`, and `ruff`.

Resolves #86